ml/requirements.txt
tensorflow-gpu
psycopg2-binary
boto3
pandas
numpy
python-dotenv
pascal-voc-writer
scikit-image
keras
memory-profiler
ffmpy
